On refait tout parce quon est dans la merde

+++ NET_Project/DLL_Library_NET_Project/Persistance/DAO/RecetteDAO.cs
@@ -6,7 +6,7 @@ namespace DLL_Library_NET_Project.Persistance.DAO
     {
         public int id { get; set; }
         public string intitule { get; set; }
-        public double prix { get; set; }
+        public int prix { get; set; }
         public int temps_prepa { get; set; }
         public int typerecetteId { get; set; }
